+++ packages/react/src/experimental/Forms/Fields/Input/components/PasswordVisibilityToggle.tsx
@@ -11,6 +11,7 @@ export const PasswordVisibilityToggle = ({
   value,
   onChange,
 }: PasswordVisibilityToggleProps) => {
+
   const i18n = useI18n()
   return (
     <F0ButtonToggle