+++ packages/react/src/experimental/OneDataCollection/OneDatacollection.tsx
@@ -45,6 +45,7 @@ import {
   RecordType,
 } from "@/hooks/datasource"
 import { useDebounceBoolean } from "@/lib/useDebounceBoolean"
+import { useDeepCompareEffect } from "@reactuses/core"
 import { NavigationFilters as NavigationFiltersComponent } from "./components/NavigationFilters"
 import { TotalItemsSummary } from "./components/TotalItemsSummary"
 import {
@@ -527,7 +528,7 @@ const OneDataCollectionComp = <
   })
 
   /** State */
-  useEffect(() => {
+  useDeepCompareEffect(() => {
     onStateChange?.({
       filters: currentFilters,
       sortings: currentSortings as SortingsState<SortingsDefinition>,
@@ -538,7 +539,6 @@ const OneDataCollectionComp = <
       settings: settings,
     })
   }, [
-    onStateChange,
     currentFilters,
     currentSearch,
     currentNavigationFilters,